+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `S3.aws_config()`: read `~/.aws/config` with typed getters.
+
 ## [0.1.0] - 2026-06-16
 
 - Initial release.